chore: renamed screen1 to ChatHomeScreen, removed unused screens
